@@ -11,30 +11,25 @@ using Microsoft.VisualStudio.Extensibility.Commands;
 [VisualStudioContribution]
 internal class SampleCommand : Command
 {
-	public SampleCommand(VisualStudioExtensibility extensibility)
-		: base(extensibility)
-	{
-	}
+    /// <inheritdoc />
+    public override CommandConfiguration CommandConfiguration => new("%CommandParentingSample.SampleCommand.DisplayName%")
+    {
+        Placements = new[]
+        {
+            // File in project context menu
+            CommandPlacement.VsctParent(new Guid("{d309f791-903f-11d0-9efc-00a0c911004f}"), id: 521, priority: 0),
 
-	/// <inheritdoc />
-	public override CommandConfiguration CommandConfiguration => new("%CommandParentingSample.SampleCommand.DisplayName%")
-	{
-		Placements = new[]
-		{
-			// File in project context menu
-			CommandPlacement.FromVsctParent(new Guid("{d309f791-903f-11d0-9efc-00a0c911004f}"), 521),
+            // Project context menu
+            CommandPlacement.VsctParent(new Guid("{d309f791-903f-11d0-9efc-00a0c911004f}"), id: 518, priority: 0),
 
-			// Project context menu
-			CommandPlacement.FromVsctParent(new Guid("{d309f791-903f-11d0-9efc-00a0c911004f}"), 518),
+            // Solution context menu
+            CommandPlacement.VsctParent(new Guid("{d309f791-903f-11d0-9efc-00a0c911004f}"), id: 537, priority: 0),
+        },
+    };
 
-			// Solution context menu
-			CommandPlacement.FromVsctParent(new Guid("{d309f791-903f-11d0-9efc-00a0c911004f}"), 537),
-		},
-	};
-
-	/// <inheritdoc />
-	public override Task ExecuteCommandAsync(IClientContext context, CancellationToken cancellationToken)
-	{
-		return Task.CompletedTask;
-	}
+    /// <inheritdoc />
+    public override Task ExecuteCommandAsync(IClientContext context, CancellationToken cancellationToken)
+    {
+        return Task.CompletedTask;
+    }
 }
