@@ -12,19 +12,20 @@ using Microsoft.VisualStudio.Extensibility;
 [VisualStudioContribution]
 public class DialogSampleExtension : Extension
 {
-	/// <inheritdoc/>
-	public override ExtensionConfiguration ExtensionConfiguration => new()
-	{
-		Metadata = new(
-				id: "DialogSampleExtension.8709efb4-fe22-4848-8533-594f444861a1",
-				version: this.ExtensionAssemblyVersion,
-				publisherName: "Microsoft",
-				displayName: "Dialog Sample Extension"),
-	};
+    /// <inheritdoc/>
+    public override ExtensionConfiguration ExtensionConfiguration => new()
+    {
+        Metadata = new(
+                id: "DialogSampleExtension.8709efb4-fe22-4848-8533-594f444861a1",
+                version: this.ExtensionAssemblyVersion,
+                publisherName: "Microsoft",
+                displayName: "Dialog Sample Extension",
+                description: "Sample extension demonstrating modal dialogs"),
+    };
 
-	/// <inheritdoc/>
-	protected override void InitializeServices(IServiceCollection serviceCollection)
-	{
-		base.InitializeServices(serviceCollection);
-	}
+    /// <inheritdoc/>
+    protected override void InitializeServices(IServiceCollection serviceCollection)
+    {
+        base.InitializeServices(serviceCollection);
+    }
 }
