@@ -17,55 +17,56 @@ using Microsoft.VisualStudio.ProjectSystem.Query;
 /// A command to execute linter on the current file selected in Solution Explorer.
 /// </summary>
 /// <remarks>
-/// This command utilizes <see cref="CommandEnabledWhenAttribute"/> to describe when commmand state is enabled.
+/// This command utilizes <see cref="CommandEnabledWhenAttribute"/> to describe when command state is enabled.
 /// </remarks>
 [VisualStudioContribution]
 internal class RunLinterOnCurrentFileCommand : Command
 {
-	private readonly TraceSource logger;
-	private MarkdownDiagnosticsService diagnosticsProvider;
+    private readonly TraceSource logger;
 
-	/// <summary>
-	/// Initializes a new instance of the <see cref="RunLinterOnCurrentFileCommand"/> class.
-	/// </summary>
-	/// <param name="extensibility">Extensibility object.</param>
-	/// <param name="traceSource">Logger instance that can be used to log extension actions.</param>
-	/// <param name="diagnosticsProvider">Local diagnostics provider service instance.</param>
-	public RunLinterOnCurrentFileCommand(VisualStudioExtensibility extensibility, TraceSource traceSource, MarkdownDiagnosticsService diagnosticsProvider)
-		: base(extensibility)
-	{
-		this.logger = Requires.NotNull(traceSource, nameof(traceSource));
-		this.diagnosticsProvider = Requires.NotNull(diagnosticsProvider, nameof(diagnosticsProvider));
+#pragma warning disable CA2213 // This is an extension scoped service.
+    private readonly MarkdownDiagnosticsService diagnosticsProvider;
+#pragma warning restore CA2213
 
-		this.logger.TraceEvent(TraceEventType.Information, 0, $"Initializing {nameof(RunLinterOnCurrentFileCommand)} instance.");
-	}
+    /// <summary>
+    /// Initializes a new instance of the <see cref="RunLinterOnCurrentFileCommand"/> class.
+    /// </summary>
+    /// <param name="traceSource">Logger instance that can be used to log extension actions.</param>
+    /// <param name="diagnosticsProvider">Local diagnostics provider service instance.</param>
+    public RunLinterOnCurrentFileCommand(TraceSource traceSource, MarkdownDiagnosticsService diagnosticsProvider)
+    {
+        this.logger = Requires.NotNull(traceSource, nameof(traceSource));
+        this.diagnosticsProvider = Requires.NotNull(diagnosticsProvider, nameof(diagnosticsProvider));
 
-	/// <inheritdoc />
-	public override CommandConfiguration CommandConfiguration => new("%MarkdownLinter.RunLinterOnCurrentFileCommand.DisplayName%")
-	{
-		Placements = new[] { CommandPlacement.KnownPlacements.ToolsMenu },
-		Icon = new(ImageMoniker.Custom("MarkdownIcon"), IconSettings.IconAndText),
-		EnabledWhen = ActivationConstraint.ClientContext(ClientContextKey.Shell.ActiveSelectionFileName, ".+"),
-	};
+        this.logger.TraceEvent(TraceEventType.Information, 0, $"Initializing {nameof(RunLinterOnCurrentFileCommand)} instance.");
+    }
 
-	/// <inheritdoc />
-	public override async Task ExecuteCommandAsync(IClientContext context, CancellationToken cancellationToken)
-	{
-		cancellationToken.ThrowIfCancellationRequested();
-		try
-		{
-			// Get the selected item URIs from IDE context that reprents the state when command was executed.
-			var selectedItemPaths = new Uri[] { await context.GetSelectedPathAsync(cancellationToken) };
+    /// <inheritdoc />
+    public override CommandConfiguration CommandConfiguration => new("%MarkdownLinter.RunLinterOnCurrentFileCommand.DisplayName%")
+    {
+        Placements = new[] { CommandPlacement.KnownPlacements.ToolsMenu },
+        Icon = new(ImageMoniker.Custom("MarkdownIcon"), IconSettings.IconAndText),
+        EnabledWhen = ActivationConstraint.ClientContext(ClientContextKey.Shell.ActiveSelectionFileName, ".+"),
+    };
 
-			// Enumerate through each selection and run linter on each selected item.
-			foreach (var selectedItem in selectedItemPaths.Where(p => p.IsFile))
-			{
-				await this.diagnosticsProvider.ProcessFileAsync(selectedItem, cancellationToken);
-			}
-		}
-		catch (InvalidOperationException ex)
-		{
-			this.logger.TraceEvent(TraceEventType.Error, 0, ex.ToString());
-		}
-	}
+    /// <inheritdoc />
+    public override async Task ExecuteCommandAsync(IClientContext context, CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+        try
+        {
+            // Get the selected item URIs from IDE context that represents the state when command was executed.
+            var selectedItemPaths = new Uri[] { await context.GetSelectedPathAsync(cancellationToken) };
+
+            // Enumerate through each selection and run linter on each selected item.
+            foreach (var selectedItem in selectedItemPaths.Where(p => p.IsFile))
+            {
+                await this.diagnosticsProvider.ProcessFileAsync(selectedItem, cancellationToken);
+            }
+        }
+        catch (InvalidOperationException ex)
+        {
+            this.logger.TraceEvent(TraceEventType.Error, 0, ex.ToString());
+        }
+    }
 }
