@@ -25,83 +25,84 @@ using Microsoft.VisualStudio.RpcContracts.ProgressReporting;
 [VisualStudioContribution]
 internal class RunLinterOnSolutionCommand : Command
 {
-	private readonly TraceSource logger;
-	private MarkdownDiagnosticsService diagnosticsProvider;
+    private readonly TraceSource logger;
 
-	/// <summary>
-	/// Initializes a new instance of the <see cref="RunLinterOnSolutionCommand"/> class.
-	/// </summary>
-	/// <param name="extensibility">Extensibility object.</param>
-	/// <param name="traceSource">Logger instance that can be used to log extension actions.</param>
-	/// <param name="diagnosticsProvider">Local diagnostics provider service instance.</param>
-	public RunLinterOnSolutionCommand(VisualStudioExtensibility extensibility, TraceSource traceSource, MarkdownDiagnosticsService diagnosticsProvider)
-		: base(extensibility)
-	{
-		this.logger = Requires.NotNull(traceSource, nameof(traceSource));
-		this.diagnosticsProvider = Requires.NotNull(diagnosticsProvider, nameof(diagnosticsProvider));
+#pragma warning disable CA2213 // This is an extension scoped service.
+    private readonly MarkdownDiagnosticsService diagnosticsProvider;
+#pragma warning restore CA2213
 
-		this.logger.TraceEvent(TraceEventType.Information, 0, $"Initializing {nameof(RunLinterOnSolutionCommand)} instance.");
-	}
+    /// <summary>
+    /// Initializes a new instance of the <see cref="RunLinterOnSolutionCommand"/> class.
+    /// </summary>
+    /// <param name="traceSource">Logger instance that can be used to log extension actions.</param>
+    /// <param name="diagnosticsProvider">Local diagnostics provider service instance.</param>
+    public RunLinterOnSolutionCommand(TraceSource traceSource, MarkdownDiagnosticsService diagnosticsProvider)
+    {
+        this.logger = Requires.NotNull(traceSource, nameof(traceSource));
+        this.diagnosticsProvider = Requires.NotNull(diagnosticsProvider, nameof(diagnosticsProvider));
 
-	/// <inheritdoc />
-	public override CommandConfiguration CommandConfiguration => new("%MarkdownLinter.RunLinterOnSolutionCommand.DisplayName%")
-	{
-		Placements = new[] { CommandPlacement.KnownPlacements.ToolsMenu },
-		Icon = new(ImageMoniker.Custom("MarkdownIcon"), IconSettings.IconAndText),
-		EnabledWhen = ActivationConstraint.SolutionState(SolutionState.FullyLoaded),
-	};
+        this.logger.TraceEvent(TraceEventType.Information, 0, $"Initializing {nameof(RunLinterOnSolutionCommand)} instance.");
+    }
 
-	/// <inheritdoc />
-	public override async Task ExecuteCommandAsync(IClientContext context, CancellationToken cancellationToken)
-	{
-		try
-		{
-			var markdownFiles = await this.Extensibility.Workspaces().QueryProjectsAsync(
-				query => query.Get(project => project.Files).With(file => file.Path).Where(file => file.Extension == ".md"),
-				cancellationToken);
+    /// <inheritdoc />
+    public override CommandConfiguration CommandConfiguration => new("%MarkdownLinter.RunLinterOnSolutionCommand.DisplayName%")
+    {
+        Placements = new[] { CommandPlacement.KnownPlacements.ToolsMenu },
+        Icon = new(ImageMoniker.Custom("MarkdownIcon"), IconSettings.IconAndText),
+        EnabledWhen = ActivationConstraint.SolutionState(SolutionState.FullyLoaded),
+    };
 
-			List<Uri> filesToProcess = new List<Uri>(markdownFiles.Select(f => new Uri(f.Path)));
-			if (filesToProcess.Count == 0)
-			{
-				return;
-			}
+    /// <inheritdoc />
+    public override async Task ExecuteCommandAsync(IClientContext context, CancellationToken cancellationToken)
+    {
+        try
+        {
+            var markdownFiles = await this.Extensibility.Workspaces().QueryProjectsAsync(
+                query => query.Get(project => project.Files).With(file => file.Path).Where(file => file.Extension == ".md"),
+                cancellationToken);
 
-			if (await this.Extensibility.Shell().ShowPromptAsync(
-				string.Format(Strings.Culture, Strings.MarkdownSolutionAnalysisPrompt, filesToProcess.Count),
-				PromptOptions.OKCancel,
-				cancellationToken))
-			{
-				await this.ProcessFilesAsync(filesToProcess, cancellationToken);
-			}
-		}
-		catch (InvalidOperationException ex)
-		{
-			this.logger.TraceEvent(TraceEventType.Error, 0, ex.ToString());
-		}
-	}
+            List<Uri> filesToProcess = new List<Uri>(markdownFiles.Select(f => new Uri(f.Path)));
+            if (filesToProcess.Count == 0)
+            {
+                return;
+            }
 
-	private static ProgressStatus CreateProgressStatus(int current, int max)
-	{
-		return new ProgressStatus((int)((current / (double)max) * 100));
-	}
+            if (await this.Extensibility.Shell().ShowPromptAsync(
+                string.Format(Strings.Culture, Strings.MarkdownSolutionAnalysisPrompt, filesToProcess.Count),
+                PromptOptions.OKCancel,
+                cancellationToken))
+            {
+                await this.ProcessFilesAsync(filesToProcess, cancellationToken);
+            }
+        }
+        catch (InvalidOperationException ex)
+        {
+            this.logger.TraceEvent(TraceEventType.Error, 0, ex.ToString());
+        }
+    }
 
-	private async System.Threading.Tasks.Task<ProgressReporter> ProcessFilesAsync(List<Uri> filesToProcess, CancellationToken cancellationToken)
-	{
-		using ProgressReporter progress = await this.Extensibility.Shell().StartProgressReportingAsync(
-			Strings.MarkdownAnalysisMessage,
-			new ProgressReporterOptions(isWorkCancellable: true),
-			cancellationToken);
+    private static ProgressStatus CreateProgressStatus(int current, int max)
+    {
+        return new ProgressStatus((int)((current / (double)max) * 100));
+    }
 
-		for (int i = 0; i < filesToProcess.Count; i++)
-		{
-			progress.CancellationToken.ThrowIfCancellationRequested();
-			progress.Report(CreateProgressStatus(i, filesToProcess.Count));
+    private async System.Threading.Tasks.Task<ProgressReporter> ProcessFilesAsync(List<Uri> filesToProcess, CancellationToken cancellationToken)
+    {
+        using ProgressReporter progress = await this.Extensibility.Shell().StartProgressReportingAsync(
+            Strings.MarkdownAnalysisMessage,
+            new ProgressReporterOptions(isWorkCancellable: true),
+            cancellationToken);
 
-			// Adding an artificial delay for demonstration purposes.
-			await Task.Delay(2000);
-			await this.diagnosticsProvider.ProcessFileAsync(filesToProcess[i], cancellationToken);
-		}
+        for (int i = 0; i < filesToProcess.Count; i++)
+        {
+            progress.CancellationToken.ThrowIfCancellationRequested();
+            progress.Report(CreateProgressStatus(i, filesToProcess.Count));
 
-		return progress;
-	}
+            // Adding an artificial delay for demonstration purposes.
+            await Task.Delay(2000);
+            await this.diagnosticsProvider.ProcessFileAsync(filesToProcess[i], cancellationToken);
+        }
+
+        return progress;
+    }
 }
