@@ -12,21 +12,22 @@ using Microsoft.VisualStudio.Extensibility;
 [VisualStudioContribution]
 internal class ExtensionEntrypoint : Extension
 {
-	/// <inheritdoc/>
-	public override ExtensionConfiguration ExtensionConfiguration => new()
-	{
-		Metadata = new(
-			id: "MemoryStreamVisualizer.97a0a2fb-f163-4fa3-91f0-48a2d4ad9f57",
-			version: this.ExtensionAssemblyVersion,
-			publisherName: "Microsoft",
-			displayName: "MemoryStream Debugger Visualizer"),
-	};
+    /// <inheritdoc/>
+    public override ExtensionConfiguration ExtensionConfiguration => new()
+    {
+        Metadata = new(
+            id: "MemoryStreamVisualizer.97a0a2fb-f163-4fa3-91f0-48a2d4ad9f57",
+            version: this.ExtensionAssemblyVersion,
+            publisherName: "Microsoft",
+            displayName: "MemoryStream Debugger Visualizer",
+            description: "A debugger visualizer for MemoryStream"),
+    };
 
-	/// <inheritdoc />
-	protected override void InitializeServices(IServiceCollection serviceCollection)
-	{
-		base.InitializeServices(serviceCollection);
+    /// <inheritdoc />
+    protected override void InitializeServices(IServiceCollection serviceCollection)
+    {
+        base.InitializeServices(serviceCollection);
 
-		// You can configure dependency injection here by adding services to the serviceCollection.
-	}
+        // You can configure dependency injection here by adding services to the serviceCollection.
+    }
 }
