@@ -13,22 +13,23 @@ using Microsoft.VisualStudio.Extensibility;
 [VisualStudioContribution]
 public class OutputWindowSampleExtension : Extension
 {
-	/// <inheritdoc/>
-	public override ExtensionConfiguration ExtensionConfiguration => new()
-	{
-		Metadata = new(
-				id: "OutputWindowSample.f739efb2-fdb2-4ecf-b857-d8e11fd1e5d3",
-				version: this.ExtensionAssemblyVersion,
-				publisherName: "Microsoft",
-				displayName: "Output Window Sample Extension"),
-	};
+    /// <inheritdoc/>
+    public override ExtensionConfiguration ExtensionConfiguration => new()
+    {
+        Metadata = new(
+                id: "OutputWindowSample.f739efb2-fdb2-4ecf-b857-d8e11fd1e5d3",
+                version: this.ExtensionAssemblyVersion,
+                publisherName: "Microsoft",
+                displayName: "Output Window Sample Extension",
+                description: "Sample extension demonstrating writing to the output window"),
+    };
 
-	/// <inheritdoc />
-	protected override ResourceManager? ResourceManager => Strings.ResourceManager;
+    /// <inheritdoc />
+    protected override ResourceManager? ResourceManager => Strings.ResourceManager;
 
-	/// <inheritdoc/>
-	protected override void InitializeServices(IServiceCollection serviceCollection)
-	{
-		base.InitializeServices(serviceCollection);
-	}
+    /// <inheritdoc/>
+    protected override void InitializeServices(IServiceCollection serviceCollection)
+    {
+        base.InitializeServices(serviceCollection);
+    }
 }
