@@ -16,25 +16,15 @@ using Microsoft.VisualStudio.RpcContracts.RemoteUI;
 [VisualStudioContribution]
 internal class RegexMatchCollectionDebuggerVisualizerProvider : DebuggerVisualizerProvider
 {
-	/// <summary>
-	/// Initializes a new instance of the <see cref="RegexMatchCollectionDebuggerVisualizerProvider"/> class.
-	/// </summary>
-	/// <param name="extension">Extension instance.</param>
-	/// <param name="extensibility">Extensibility object.</param>
-	public RegexMatchCollectionDebuggerVisualizerProvider(RegexMatchVisualizerExtension extension, VisualStudioExtensibility extensibility)
-		: base(extension, extensibility)
-	{
-	}
+    /// <inheritdoc/>
+    public override DebuggerVisualizerProviderConfiguration DebuggerVisualizerProviderConfiguration => new("%RegexMatchVisualizer.RegexMatchCollectionDebuggerVisualizerProvider.DisplayName%", typeof(MatchCollection))
+    {
+        VisualizerObjectSourceType = new("RegexMatchVisualizer.ObjectSource.RegexMatchCollectionObjectSource, RegexMatchObjectSource"),
+    };
 
-	/// <inheritdoc/>
-	public override DebuggerVisualizerProviderConfiguration DebuggerVisualizerProviderConfiguration => new("Regex Match visualizer", typeof(MatchCollection))
-	{
-		VisualizerObjectSourceType = new("RegexMatchVisualizer.ObjectSource.RegexMatchCollectionObjectSource, RegexMatchObjectSource"),
-	};
-
-	/// <inheritdoc/>
-	public override Task<IRemoteUserControl> CreateVisualizerAsync(VisualizerTarget visualizerTarget, CancellationToken cancellationToken)
-	{
-		return Task.FromResult<IRemoteUserControl>(new RegexMatchCollectionVisualizerUserControl(visualizerTarget));
-	}
+    /// <inheritdoc/>
+    public override Task<IRemoteUserControl> CreateVisualizerAsync(VisualizerTarget visualizerTarget, CancellationToken cancellationToken)
+    {
+        return Task.FromResult<IRemoteUserControl>(new RegexMatchCollectionVisualizerUserControl(visualizerTarget));
+    }
 }
