@@ -4,6 +4,7 @@
 namespace RegexMatchVisualizer.ObjectSource;
 
 using System;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
@@ -14,49 +15,49 @@ using Microsoft.VisualStudio.DebuggerVisualizers;
 /// </summary>
 public class RegexMatchObjectSource : VisualizerObjectSource
 {
-	private static readonly Func<Group, string?>? GetGroupName =
-		(Func<Group, string?>?)typeof(Group).GetProperty("Name")?.GetGetMethod().CreateDelegate(typeof(Func<Group, string?>));
+    private static readonly Func<Group, string?>? GetGroupName =
+        (Func<Group, string?>?)typeof(Group).GetProperty("Name")?.GetGetMethod().CreateDelegate(typeof(Func<Group, string?>));
 
-	/// <inheritdoc/>
-	public override void GetData(object target, Stream outgoingData)
-	{
-		if (target is Match match)
-		{
-			var result = Convert(match);
-			SerializeAsJson(outgoingData, result);
-		}
-	}
+    /// <inheritdoc/>
+    public override void GetData(object target, Stream outgoingData)
+    {
+        if (target is Match match)
+        {
+            var result = Convert(match);
+            SerializeAsJson(outgoingData, result);
+        }
+    }
 
-	internal static RegexMatch Convert(Match match) =>
-		new RegexMatch
-		{
-			Success = match.Success,
-			Index = match.Index,
-			Length = match.Length,
-			Value = match.Value,
-			Captures = Convert(match.Captures),
-			Groups = Convert(match.Groups),
-		};
+    internal static RegexMatch Convert(Match match) =>
+        new RegexMatch
+        {
+            Success = match.Success,
+            Index = match.Index,
+            Length = match.Length,
+            Value = match.Value,
+            Captures = Convert(match.Captures),
+            Groups = Convert(match.Groups),
+        };
 
-	private static RegexCapture[] Convert(CaptureCollection captures) =>
-		captures.OfType<Capture>()
-			.Select((c, i) => new RegexCapture
-			{
-				Index = c.Index,
-				Length = c.Length,
-				Value = c.Value,
-				Name = $"[{i}]",
-			}).ToArray();
+    private static RegexCapture[] Convert(CaptureCollection captures) =>
+        captures.OfType<Capture>()
+            .Select((c, i) => new RegexCapture
+            {
+                Index = c.Index,
+                Length = c.Length,
+                Value = c.Value,
+                Name = $"[{i}]",
+            }).ToArray();
 
-	private static RegexGroup[] Convert(GroupCollection groups) =>
-		groups.OfType<Group>()
-			.Select((g, i) => new RegexGroup
-			{
-				Success = g.Success,
-				Index = g.Index,
-				Length = g.Length,
-				Value = g.Value,
-				Name = $"[{GetGroupName?.Invoke(g) ?? i.ToString()}]",
-				Captures = Convert(g.Captures),
-			}).ToArray();
+    private static RegexGroup[] Convert(GroupCollection groups) =>
+        groups.OfType<Group>()
+            .Select((g, i) => new RegexGroup
+            {
+                Success = g.Success,
+                Index = g.Index,
+                Length = g.Length,
+                Value = g.Value,
+                Name = $"[{GetGroupName?.Invoke(g) ?? i.ToString(CultureInfo.CurrentCulture)}]",
+                Captures = Convert(g.Captures),
+            }).ToArray();
 }
