@@ -15,37 +15,33 @@ using Microsoft.VisualStudio.RpcContracts.RemoteUI;
 [VisualStudioContribution]
 public class MyToolWindow : ToolWindow
 {
-	private object? dataContext;
+    private object? dataContext;
 
-	/// <summary>
-	/// Initializes a new instance of the <see cref="MyToolWindow" /> class.
-	/// </summary>
-	/// <param name="extensibility">
-	/// Extensibility object instance.
-	/// </param>
-	public MyToolWindow(VisualStudioExtensibility extensibility)
-		: base(extensibility)
-	{
-		this.Title = "My Tool Window";
-	}
+    /// <summary>
+    /// Initializes a new instance of the <see cref="MyToolWindow" /> class.
+    /// </summary>
+    public MyToolWindow()
+    {
+        this.Title = "My Tool Window";
+    }
 
-	/// <inheritdoc />
-	public override ToolWindowConfiguration ToolWindowConfiguration => new()
-	{
-		Placement = ToolWindowPlacement.DocumentWell,
-	};
+    /// <inheritdoc />
+    public override ToolWindowConfiguration ToolWindowConfiguration => new()
+    {
+        Placement = ToolWindowPlacement.DocumentWell,
+    };
 
-	/// <inheritdoc />
-	public override Task InitializeAsync(CancellationToken cancellationToken)
-	{
-		this.dataContext = new MyToolWindowData(this.Extensibility);
+    /// <inheritdoc />
+    public override Task InitializeAsync(CancellationToken cancellationToken)
+    {
+        this.dataContext = new MyToolWindowData(this.Extensibility);
 
-		return Task.CompletedTask;
-	}
+        return Task.CompletedTask;
+    }
 
-	/// <inheritdoc />
-	public override Task<IRemoteUserControl> GetContentAsync(CancellationToken cancellationToken)
-	{
-		return Task.FromResult<IRemoteUserControl>(new MyToolWindowControl(this.dataContext));
-	}
+    /// <inheritdoc />
+    public override Task<IRemoteUserControl> GetContentAsync(CancellationToken cancellationToken)
+    {
+        return Task.FromResult<IRemoteUserControl>(new MyToolWindowControl(this.dataContext));
+    }
 }
