@@ -13,19 +13,20 @@ using Microsoft.VisualStudio.Extensibility;
 [VisualStudioContribution]
 internal class WordCountMarginSampleExtension : Extension
 {
-	/// <inheritdoc/>
-	public override ExtensionConfiguration ExtensionConfiguration => new()
-	{
-		Metadata = new(
-				id: "WordCountMarginSample.9d845a9e-a3a9-4efe-97f9-002a3945f9f5",
-				version: this.ExtensionAssemblyVersion,
-				publisherName: "Microsoft",
-				displayName: "Word Count Margin Sample Extension"),
-	};
+    /// <inheritdoc/>
+    public override ExtensionConfiguration ExtensionConfiguration => new()
+    {
+        Metadata = new(
+                id: "WordCountMarginSample.9d845a9e-a3a9-4efe-97f9-002a3945f9f5",
+                version: this.ExtensionAssemblyVersion,
+                publisherName: "Microsoft",
+                displayName: "Word Count Margin Sample Extension",
+                description: "Sample extension demonstrating custom margins"),
+    };
 
-	/// <inheritdoc/>
-	protected override void InitializeServices(IServiceCollection serviceCollection)
-	{
-		base.InitializeServices(serviceCollection);
-	}
+    /// <inheritdoc/>
+    protected override void InitializeServices(IServiceCollection serviceCollection)
+    {
+        base.InitializeServices(serviceCollection);
+    }
 }
